% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BB_Sky_API_Base_Func.R
\name{bb_get_func}
\alias{bb_get_func}
\title{Title}
\usage{
bb_get_func(
  api_url = bb_url_list$school,
  url1 = NULL,
  url2 = NULL,
  url3 = NULL,
  url4 = NULL,
  url5 = NULL,
  limit = 500,
  offset = 0,
  page = 1,
  marker = 0,
  params = NULL,
  testing = F,
  no_format = F,
  no_page = F,
  method = "GET",
  body = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
